Remove upload public key button

+++ encryptFile.py
@@ -34,6 +34,7 @@ class EncryptFile():
 		self.__splitFile()
 		self.__encryptChunks()
 		self.__populateFilesAndManifest()
+		self.__uploadFilesToGoogleDrive()
 		return self.__ctChunks
 
 	#Private method to read the file		
@@ -93,9 +94,9 @@ class EncryptFile():
 		filesWrite.write("\n")
 		filesWrite.close()
 
-		gauth = GoogleAuth()
-		gauth.LocalWebserverAuth()
-		drive = GoogleDrive(gauth)
+		#gauth = GoogleAuth()
+		#gauth.LocalWebserverAuth()
+		#drive = GoogleDrive(gauth)
 		
 		for chunk in self.__ctChunks:
 			salt = base64.urlsafe_b64encode(uuid.uuid4().bytes)		
@@ -120,11 +121,32 @@ class EncryptFile():
 					
 			manifestWrite.write(hashObj.hexdigest() + "\t" + salt + "\n")
 
-			file = drive.CreateFile({'title': encFileName})
-			file.SetContentFile(encFilePath)
-			file.Upload()
+			#file = drive.CreateFile({'title': encFileName})
+			#file.SetContentFile(encFilePath)
+			#file.Upload()
 
-			os.remove(encFilePath)
+			#os.remove(encFilePath)
 			
 		manifestWrite.write("\n")
 		manifestWrite.close()
+
+	#Method to upload files to Google Drive
+	def __uploadFilesToGoogleDrive(self):
+		workingDirectory = os.getcwd()
+		storeDirectory = os.path.join(workingDirectory, "Store")
+		
+		encFiles = os.listdir(storeDirectory)
+
+		gauth = GoogleAuth()
+		gauth.LocalWebserverAuth()
+		drive = GoogleDrive(gauth)
+				
+		for encFileName in encFiles:
+			encFilePath = os.path.join(storeDirectory, encFileName)
+			if os.path.isfile(encFilePath):
+				uploadFile = drive.CreateFile({"title": encFileName})
+				uploadFile.SetContentFile(encFilePath)
+				uploadFile.Upload()
+				os.remove(encFilePath)
+						
+
